refactor: separate regex parsing logic
