+++ src/plugins/flow/FlowTypePlugin.js
@@ -1,19 +1,8 @@
-export function FlowTypeAnnotationReader(stream, factory) {
-  const start = stream.getPosition();
-  if (stream.current() !== ':') return null;
-  let value = ':';
-  stream.advance();
-  while (stream.current() && /\s/.test(stream.current())) {
-    value += stream.current();
-    stream.advance();
-  }
-  while (stream.current() && /[A-Za-z0-9_<>,\s?]/.test(stream.current())) {
-    value += stream.current();
-    stream.advance();
-    if (/\n/.test(stream.current())) break;
-  }
-  return factory('TYPE_ANNOTATION', value.trimEnd(), start, stream.getPosition());
-}
+import { createTypeAnnotationReader } from '../common/TypeAnnotationReader.js';
+
+export const FlowTypeAnnotationReader = createTypeAnnotationReader({
+  allowQuestionMark: true
+});
 
 export const FlowTypePlugin = {
   modes: { default: [FlowTypeAnnotationReader] },
